+++ client/package.json
@@ -14,7 +14,8 @@
     "react-tap-event-plugin": "^2.0.1",
     "redux": "^3.6.0",
     "redux-form": "^6.6.3",
-    "socket.io-client": "^1.7.3"
+    "socket.io-client": "^1.7.3",
+    "uuid": "^3.0.1"
   },
   "devDependencies": {
     "react-scripts": "0.9.5"
