+++ client/src/App.js
@@ -2,13 +2,13 @@ import React, { Component } from 'react'
 import './App.css'
 import {connect} from 'react-redux'
 import FilterPanel from './CharlesComponents/FilterPanel'
-import FilterChips from './CharlesComponents/FilterChips'
+// import FilterChips from './CharlesComponents/FilterChips'
 
 class App extends Component {
  	  render() {
       return (
         <div>
-        	<FilterChips />
+        	
           <FilterPanel />
         </div>
       )
@@ -24,3 +24,4 @@ const mapStateToProps = function(appState) {
 }
 
 export default connect(mapStateToProps)(App)
+// <FilterChips />
