+++ client/src/CharlesComponents/Chip.js
@@ -1,0 +1,16 @@
+import React from 'react'
+import FilterForm from './FilterForm'
+
+class Chip extends React.Component {
+  constructor(props) {
+    super(props)
+  }
+  render() {
+    return (
+      <div>
+        
+      </div>
+    )
+  }
+}
+export default Chip
